% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_partialpro.R
\name{bart.learner}
\alias{bart.learner}
\title{BART learner wrapper for varPro}
\usage{
bart.learner(o, ...)
}
\arguments{
\item{o}{A varpro object.}

\item{...}{Additional learner options.}
}
\value{
A prediction function.
}
\description{
Constructs a prediction function based on BART for use in partialpro flows.
}
