% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_partialpro.R
\name{gbm.learner}
\alias{gbm.learner}
\title{GBM learner wrapper for varPro}
\usage{
gbm.learner(o, ...)
}
\arguments{
\item{o}{A varpro object.}

\item{...}{Additional GBM options.}
}
\value{
A prediction function.
}
\description{
Constructs a gradient-boosting prediction function for varpro analysis.
}
