% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_cv_varpro.R
\name{get.mc.cores}
\alias{get.mc.cores}
\title{Determine default mc.cores setting}
\usage{
get.mc.cores()
}
\value{
Integer number of cores.
}
\description{
Returns the core count used by varPro helper code, honoring user and check
environment settings.
}
