% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_importance.R
\name{get.orgvimp}
\alias{get.orgvimp}
\title{Map importance back to original variables}
\usage{
get.orgvimp(o, pretty = TRUE, local.std = TRUE, vmp = NULL)
}
\arguments{
\item{o}{A varpro, cv.varpro, or uvarpro object.}

\item{pretty}{Logical; return prettified output.}

\item{local.std}{Logical; use local standardized importance for varpro/uvarpro.}

\item{vmp}{Optional precomputed importance table.}
}
\value{
Importance summaries on original variable names.
}
\description{
Map importance back to original variables
}
