% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_importance.R
\name{get.topvars}
\alias{get.topvars}
\title{Get top selected variable names}
\usage{
get.topvars(o, local.std = TRUE)
}
\arguments{
\item{o}{A varpro or uvarpro object.}

\item{local.std}{Logical; use local standardized importance.}
}
\value{
Character vector of selected variable names.
}
\description{
Get top selected variable names
}
