% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_importance.R
\name{get.vimp}
\alias{get.vimp}
\title{Extract variable-importance values}
\usage{
get.vimp(o, pretty = TRUE, local.std = TRUE)
}
\arguments{
\item{o}{A varpro, cv.varpro, or uvarpro object.}

\item{pretty}{Logical; return compact named output.}

\item{local.std}{Logical; use local standardized importance.}
}
\value{
Importance values in named vector/list/data-frame form.
}
\description{
Extract variable-importance values
}
