% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.uvarpro.R
\name{predict.uvarpro}
\alias{predict.uvarpro}
\title{Predict from uvarpro}
\usage{
\method{predict}{uvarpro}(object, newdata, ...)
}
\arguments{
\item{object}{A uvarpro object.}

\item{newdata}{Optional prediction data.}

\item{...}{Additional arguments passed to forest prediction.}
}
\value{
Predicted matrix with mse attributes.
}
\description{
Produces reconstructed values and error summaries for uvarpro objects.
}
