% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.varpro.R
\name{predict.varpro}
\alias{predict.varpro}
\title{Predict from varpro}
\usage{
\method{predict}{varpro}(object, newdata, ...)
}
\arguments{
\item{object}{A varpro object.}

\item{newdata}{Optional prediction data.}

\item{...}{Additional arguments passed to forest prediction.}
}
\value{
Predicted values.
}
\description{
Generates predictions from a fitted varpro object.
}
