% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.varpro.R
\name{print.cv.varpro}
\alias{print.cv.varpro}
\title{Print cv.varpro}
\usage{
\method{print}{cv.varpro}(x, ...)
}
\arguments{
\item{x}{A cv.varpro object.}

\item{...}{Additional print options.}
}
\value{
The input object invisibly.
}
\description{
Print cv.varpro
}
