% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivarpro.R
\name{print.ivarpro}
\alias{print.ivarpro}
\title{Print ivarpro}
\usage{
\method{print}{ivarpro}(x, ..., full = FALSE)
}
\arguments{
\item{x}{An ivarpro object.}

\item{...}{Additional print options.}

\item{full}{Logical; print full details.}
}
\value{
The input object invisibly.
}
\description{
Print ivarpro
}
