% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_partialpro.R
\name{rf.learner}
\alias{rf.learner}
\title{Random-forest learner wrapper for varPro}
\usage{
rf.learner(o, ...)
}
\arguments{
\item{o}{A varpro object.}

\item{...}{Additional randomForestSRC options.}
}
\value{
A prediction function.
}
\description{
Constructs a randomForestSRC prediction function for varpro analysis.
}
