% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{roughfix}
\alias{roughfix}
\title{Rough imputation for mixed data}
\usage{
roughfix(data)
}
\arguments{
\item{data}{Data frame with missing values.}
}
\value{
Data frame with missing values imputed.
}
\description{
Simple rough-fix style imputation for numeric/factor data frame columns.
}
