\name{internal.utils}

\alias{bart.learner}
\alias{beta.varpro}

\alias{gbm.learner}


\alias{get.iso.performance}
\alias{get.mc.cores}
\alias{get.orgvimp}
\alias{get.splitweight.custom}
\alias{get.topvars}
\alias{get.varpro.strengthArray}
\alias{get.vimp}

\alias{importance.varpro.workhorse}

\alias{local.importance}


\alias{rf.learner}

\alias{roughfix}





\title{Internal Utility Functions}

\description{
  These are internal utility functions exported for advanced usage. This
  Rd file is used solely to register aliases.
}

\value{No return value, called for side effects.}



\keyword{internal}
