% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{varPro-package}
\alias{varPro}
\alias{varPro-package}
\title{varPro package}
\description{
Model-independent variable selection via rule-based variable priority.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.varprotools.org/ https://www.luminwin.net/ https://ishwaran.org/}
  \item Report bugs at \url{https://github.com/kogalur/varPro/issues/}
}

}
\author{
\strong{Maintainer}: Udaya B. Kogalur \email{ubk@kogalur.com}

Authors:
\itemize{
  \item Min Lu \email{luminwin@gmail.com}
  \item Aster K. Shear \email{aster.shear@gmail.com}
  \item Hemant Ishwaran \email{hemant.ishwaran@gmail.com}
}

}
\keyword{internal}
