\name{varpro.news}
\alias{varpro.news}
\title{Show the NEWS file}
\description{
  Show the NEWS file of the \pkg{varPro} package.
}
\usage{
varpro.news(...)
}
\arguments{
  \item{...}{Further arguments passed to or from other methods.}
}
\value{
  None.
}
\author{
  Hemant Ishwaran and Udaya B. Kogalur
}
\keyword{documentation}
