% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance-generic.R
\name{importance}
\alias{importance}
\title{Importance generic}
\usage{
importance(object, ...)
}
\arguments{
\item{object}{Object to evaluate.}

\item{...}{Additional arguments.}
}
\value{
Method-specific importance output.
}
\description{
Generic for importance methods.
}
