\name{internal.utils}

\alias{bart.learner}
\alias{beta.varpro}

\alias{get.beta.entropy}
\alias{get.iso.performance}
\alias{get.orgvimp}
\alias{get.topvars}
\alias{get.vimp}

\alias{gbm.learner}

\alias{make.vt}

\alias{rf.learner}

\alias{sdependent}




\title{Internal Utility Functions}

\description{
  These are internal utility functions exported for advanced usage. This
  Rd file is used solely to register aliases.
}

\keyword{internal}
